feat: removida arquitetura de monorepo
