--- core/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@tray-storefront/core",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./dist/index.js",
-  "source": "./dist",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "dev": "tsc --watch --sourcemap",
-    "build": "tsc",
-    "test": "npx jest"
-  },
-  "dependencies": {
-    "@urql/core": "^2.3.3",
-    "graphql": "^15.6.1",
-    "socket.io-client": "^4.2.0",
-    "webfontloader": "^1.6.28"
-  },
-  "devDependencies": {
-    "@types/jest": "^27.0.3",
-    "@types/node": "^16.11.6",
-    "@types/webfontloader": "^1.6.33",
-    "jest": "^27.4.2",
-    "ts-jest": "^27.0.7",
-    "typescript": "^4.4.4"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/DoocaCommerce/tray-storefront-vue.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/DoocaCommerce/tray-storefront-vue/issues"
-  },
-  "homepage": "https://github.com/DoocaCommerce/tray-storefront-vue#readme"
-}
