--- core/src/main.ts
@@ -1,1 +1,0 @@
-console.log('hello word!!!')
