+++ package.json
@@ -1,25 +1,36 @@
 {
-  "name": "@tray-storefront/react",
+  "name": "tray-storefront",
   "version": "1.0.0",
   "description": "",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "./lib/index.js",
+  "source": "./lib",
+  "types": "./lib/index.d.ts",
   "scripts": {
     "dev": "tsc --watch --sourcemap",
-    "build": "tsc"
+    "build": "tsc",
+    "test": "npx jest"
   },
   "dependencies": {
+    "@urql/core": "^2.3.3",
+    "graphql": "^15.6.1",
+    "socket.io-client": "^4.2.0",
+    "webfontloader": "^1.6.28",
     "@reduxjs/toolkit": "^1.6.2",
     "react-redux": "^7.2.6",
     "react-router": "^6.0.2",
     "react-router-dom": "^6.0.2"
   },
   "devDependencies": {
+    "@types/jest": "^27.0.3",
+    "@types/node": "^16.11.6",
+    "@types/webfontloader": "^1.6.33",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "typescript": "^4.3.2"
+    "jest": "^27.4.2",
+    "ts-jest": "^27.0.7",
+    "typescript": "^4.4.4"
   },
   "peerDependencies": {
     "react": "^17.0.2",
