--- react/src/@types/RootState.ts
@@ -1,8 +1,0 @@
-import { SectionsState } from 'src/store/modules/sections/type'
-import { SettingsState } from 'src/store/modules/settings/type'
-import { CartResumeState } from '../store/modules/cart-resume/type'
-export interface RootState {
-  settings: SettingsState
-  sections: SectionsState
-  cartResume: CartResumeState
-}
