--- react/src/store/modules/sections/type.ts
@@ -1,3 +1,0 @@
-import { BaseState } from '../../../@types/BaseState'
-
-export interface SectionsState extends BaseState<Object> {}
