--- react/src/store/modules/settings/type.ts
@@ -1,3 +1,0 @@
-import { BaseState } from '../../../@types/BaseState'
-
-export interface SettingsState extends BaseState<Object> {}
