--- react/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "module": "esnext",
-    "moduleResolution": "node",
-    "allowJs": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "resolveJsonModule": true,
-    "allowSyntheticDefaultImports": true,
-    "declaration": true,
-    "sourceMap": true,
-    "outDir": "dist/",
-    "lib": ["dom", "esnext"],
-    "jsx": "react"
-  },
-  "include": ["src/"]
-}
