+++ src/common/types/ImageTypes.ts
