+++ src/core/helpers/stringHelper.ts
