+++ src/core/helpers/styleHelper.ts
