+++ src/core/socket/updateSettingsData.ts
