+++ src/react/store/modules/sections/type.ts
@@ -1,0 +1,3 @@
+import { BaseState } from '../../../../common/types/BaseState'
+
+export interface SectionsState extends BaseState<Object> {}
