+++ src/react/store/modules/settings/type.ts
@@ -1,0 +1,3 @@
+import { BaseState } from '../../../../common/types/BaseState'
+
+export interface SettingsState extends BaseState<Object> {}
