+++ src/vue/components/RootVars.vue
