+++ src/vue/components/SectionLoader.vue
